requirements.txt
requests~=2.28.1
Django~=4.1.1
PyYAML~=6.0
dateparser~=1.1.1
parameterized~=0.8.1
python-dateutil~=2.8.2